refactor: move schema setup to config folder
