+++ apps/server/src/config/schema.ts
@@ -1,5 +1,5 @@
 import { buildSchema } from "type-graphql";
-import { HelloResolver } from "./modules/HelloResolver";
+import { HelloResolver } from "../modules/HelloResolver";
 
 export const createSchema = async () =>
   buildSchema({
